# perf(metrics): Use kafka header optimization