Use kafka headers to determine whether a message is applicable for the specific consumer. This was an optimization present in the python based consumers. Bringing it to the rust based consumer.

## Benchmarking results
### Counters
**Without optimization**
time:   [680.71 µs 692.80 µs 706.24 µs]
**With optimization**
time:   [2.7085 µs 2.7586 µs 2.8125 µs]

### Distributions
**Without optimization**
time:   [1.4514 ms 1.4923 ms 1.5335 ms]
**With optimization**
time:   [1.0093 ms 1.0364 ms 1.0636 ms]

## Tests
Unit tests have been added.

Also did end to end testing by making the following modification to the code and running `bin/send_metrics.py` from sentry
```
diff --git a/rust_snuba/src/processors/generic_metrics.rs b/rust_snuba/src/processors/generic_metrics.rs
index 522fbb9bb..3333f3ef3 100644
--- a/rust_snuba/src/processors/generic_metrics.rs
+++ b/rust_snuba/src/processors/generic_metrics.rs
@@ -350,7 +350,10 @@ pub fn process_counter_message(
         MetricTypeHeader::Counter | MetricTypeHeader::Unknown => {
             process_message::<CountersRawRow>(payload, config)
         }
-        _ => Ok(InsertBatch::skip()),
+        _ => {
+            tracing::info!("Skipping non counter message");
+            Ok(InsertBatch::skip())
+        },
     }
 }

@@ -427,7 +430,10 @@ pub fn process_set_message(
         MetricTypeHeader::Set | MetricTypeHeader::Unknown => {
             process_message::<SetsRawRow>(payload, config)
         }
-        _ => Ok(InsertBatch::skip()),
+        _ => {
+            tracing::info!("Skipping non set message");
+            Ok(InsertBatch::skip())
+        },
     }
 }

@@ -510,7 +516,10 @@ pub fn process_distribution_message(
         MetricTypeHeader::Distribution | MetricTypeHeader::Unknown => {
             process_message::<DistributionsRawRow>(payload, config)
         }
-        _ => Ok(InsertBatch::skip()),
+        _ => {
+            tracing::info!("Skipping non distribution message");
+            Ok(InsertBatch::skip())
+        },
     }
 }

@@ -612,7 +621,10 @@ pub fn process_gauge_message(
         MetricTypeHeader::Gauge | MetricTypeHeader::Unknown => {
             process_message::<GaugesRawRow>(payload, config)
         }
-        _ => Ok(InsertBatch::skip()),
+        _ => {
+            tracing::info!("Skipping non gauge message");
+            Ok(InsertBatch::skip())
+        },
     }
 }
```

In the logs you can see the `Skipping non XXX message`
```
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.499916Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.500361Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.500620Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.500887Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.501176Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.501449Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.501746Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.502056Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.502357Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.502642Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.502885Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.503114Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.503523Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.503822Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.504309Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.504605Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-sets-consumer                 | {"timestamp":"2024-04-11T18:39:51.516838Z","level":"INFO","fields":{"message":"Skipping non set message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-sets-consumer                 | {"timestamp":"2024-04-11T18:39:51.517206Z","level":"INFO","fields":{"message":"Skipping non set message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-sets-consumer                 | {"timestamp":"2024-04-11T18:39:51.517504Z","level":"INFO","fields":{"message":"Skipping non set message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-sets-consumer                 | {"timestamp":"2024-04-11T18:39:51.518581Z","level":"INFO","fields":{"message":"Skipping non set message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.516966Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-sets-consumer                 | {"timestamp":"2024-04-11T18:39:51.519094Z","level":"INFO","fields":{"message":"Skipping non set message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.552595Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.553985Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.554291Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.555443Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.555696Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.556017Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.556303Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.556665Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.556994Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.557239Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.557498Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.557749Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.557998Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.558218Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.558462Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.558679Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.558887Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.559419Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.559638Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.559886Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-distributions-consumer        | {"timestamp":"2024-04-11T18:39:51.560131Z","level":"INFO","fields":{"message":"Skipping non distribution message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.561419Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.561689Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.561950Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.562206Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-gauges-consumer               | {"timestamp":"2024-04-11T18:39:51.563185Z","level":"INFO","fields":{"message":"Skipping non gauge message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.563573Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.563876Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.564154Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.564388Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.564605Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.564856Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-sets-consumer                 | {"timestamp":"2024-04-11T18:39:51.565020Z","level":"INFO","fields":{"message":"Skipping non set message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.565121Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-sets-consumer                 | {"timestamp":"2024-04-11T18:39:51.565306Z","level":"INFO","fields":{"message":"Skipping non set message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-sets-consumer                 | {"timestamp":"2024-04-11T18:39:51.565558Z","level":"INFO","fields":{"message":"Skipping non set message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
11:39:51 generic-metrics-counters-consumer             | {"timestamp":"2024-04-11T18:39:51.565674Z","level":"INFO","fields":{"message":"Skipping non counter message"},"target":"rust_snuba::processors::generic_metrics","span":{"name":"process_payload"},"spans":[{"name":"process_payload"}]}
…